Remove HybridMethod descriptor
This feature was used to treat the Configuration object like a global
singleton. This ended up creating a more confusing interface. It also
made it difficult to create subclasses, for testing or other purposes,
which are not affected by changes in the parent.

Instead, each instance of Configuration will now share a class dict
which stores the namespaces.

Signed-off-by: Kevin Conway <kevinjacobconway@gmail.com>