Add broader API for Namespace objects
The new features allow for an alternative way of interactive with
the Namespace objects in addition to the simple attribute access
for getting and setting option values. The option objects can now
be iterated over in addition to being resolved like descriptors.

The descriptor module is no longer used for Namespaces and has
been removed.

Signed-off-by: Kevin Conway <kevinjacobconway@gmail.com>