Swap readfp with read_file for method name
At least one version of pylint detects any call to ConfigParser.readfp
as deprecated even with the compatibility shim. This swaps the method
names so that read_file is the default and readfp is a legacy fallback.
It both satisfies the linter and more correctly stages the code to have
the shim removed later.